# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stadion Tracker API",
        "description": "API-documentatie voor stadionbezoeken, wedstrijden, posts en gebruikersinteractie.",
        "contact": {
            "email": "cedricvanhoorebeke@outlook.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://backend.ddev.site",
            "description": "DDEV ontwikkelomgeving"
        }
    ],
    "paths": {
...
        "/api/teams": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Toon alle teams",
                "operationId": "fc1c5dd9ef530faf6e8bba930cad296c",
                "responses": {
                    "200": {
                        "description": "Lijst van teams",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Team"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Voeg een nieuw team toe",
                "operationId": "b05c08b4a03b1d708fc94bf7dd5f8e46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "RSC Anderlecht"
                                    },
                                    "league_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "profile_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "banner_image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Team aangemaakt",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Team"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validatiefout"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/teams/{id}": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Toon een specifiek team",
                "operationId": "ab05003decb5cbc16154382836f51e3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Team details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Team"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Team niet gevonden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Update teamgegevens (PATCH via _method)",
                "operationId": "182759eb9561676d895be2816abda575",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PATCH"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Club Brugge"
                                    },
                                    "league_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "profile_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "banner_image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Team geüpdatet",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Team"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validatiefout"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Teams"
                ],
                "summary": "Verwijder een team",
                "operationId": "9b992594932c3e54aa58b58c3c1e933c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Team verwijderd"
                    },
                    "404": {
                        "description": "Team niet gevonden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Team": {
                "required": [
                    "name",
                    "league_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Manchester City"
                    },
                    "league_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "profile_image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://your-bucket.s3.region.amazonaws.com/teams/profile-image/abc.jpg",
                        "nullable": true
                    },
                    "banner_image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://your-bucket.s3.region.amazonaws.com/teams/banner-image/xyz.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-01T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-01T10:30:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}